third_party/WebKit/LayoutTests/fast/events/touch/gesture/gesture-tap-hit-test-counts.html

<!DOCTYPE html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<style>
html {
  font-family: Ahem;
  font-size: 10px;
}
#testArea {
  position: absolute;
  right: 50px;
  top: 50px;
}
#target {
  width: 10px;
  height: 10px;
}
#frame {
  width: 100px;
  height: 100px;
  margin-top: 30px;
}
</style>
<div id=testArea>
  <div id=target></div>
  <iframe id=frame srcdoc='<body style="margin: 5px"><iframe width=75 height=75></iframe></body>'></iframe>
</div>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<script>
const expectedHittestCount = 9;

const expectedEvents = ['pointerdown', 'touchstart', 'pointerup', 'touchend', 'mousemove', 'mousedown', 'mouseup', 'click'];
const eventLogRecorder = [];
for (eventName of expectedEvents) {
    document.addEventListener(eventName, function(event){
        eventLogRecorder.push(`${event.type}`);
    });
}

function hitTestCountDelta(doc)
{
    var lastCount = 0;
    if ('lastHitTestCount' in doc)
      lastCount = doc.lastHitTestCount;
    var newCount = internals.hitTestCount(doc);
    doc.lastHitTestCount = newCount;
    return newCount - lastCount;
}

function hitTestCacheHitsDelta(doc)
{
    var lastCount = 0;
    if ('lastHitTestCacheHits' in doc)
      lastCount = doc.lastHitTestCacheHits;
    var newCount = internals.hitTestCacheHits(doc);
    doc.lastHitTestCacheHits = newCount;
    return newCount - lastCount;
}

function clearCounts(documents)
{
    for(var i = 0; i < documents.length; i++) {
        documents[i].lastHitTestCount = internals.hitTestCount(documents[i]);
        documents[i].lastHitTestCacheHits = internals.hitTestCacheHits(documents[i]);
        internals.clearHitTestCache(documents[i]);
    }
}

function centerOf(element) {
    var targetRect = element.getBoundingClientRect();
    return {
      x: targetRect.left + targetRect.width / 2,
      y: targetRect.top + targetRect.height / 2
    };
}

function sendTapAt(targetX, targetY) {
    return new Promise(function(resolve, reject) {
        if (window.chrome && chrome.gpuBenchmarking) {
          var pointerActions =
              [{source: "touch",
                actions: [
                  { name: "pointerDown", x: targetX, y: targetY },
                  { name: "pointerUp" },
              ]}];
          chrome.gpuBenchmarking.pointerActionSequence(pointerActions, resolve);
        }
        else {
            reject();
        }
    });
}


promise_test(async () => {
    clearCounts([document]);
    var point = centerOf(document.getElementById('target'));
    await sendTapAt(point.x, point.y);

    assert_array_equals(eventLogRecorder, expectedEvents);
    var countStr = '';
    var hits = hitTestCountDelta(document);
    assert_equals(hits, expectedHittestCount);
}, "simple tap");

</script>
